Edited which files are deleted at end
Changed the method in which files are deleted after the creation of dta
datasets so that only the folders created during the process are
checked. Keeping the unzipped codebooks files instead of the zipped
codebooks files because they will be referenced often. Added a line of
code to the create dta function that creates the clean data folder if
it doesn’t already exist.

+++ data_scrape/delete_extracted_files.py
@@ -15,38 +15,32 @@ ipeds_directory = os.path.join(os.path.dirname(os.path.abspath(__file__)), os.pa
 # Go through all the files and folders contained in the directory, now including the created dta files
 for directory_path, directory_names_list, file_names_list in os.walk(ipeds_directory):
     directory_name = os.path.basename(directory_path)
-    if re.fullmatch("[0-9]{4}", directory_name):
-        for file_name in file_names_list:
-            file_path = os.path.join(directory_path, file_name)
-            # Check to see if the file came from a zipped file
-            print("Checking if " + file_name + " is a do, csv, html, or codebook zip file")
-
-            if file_name.endswith(".dta"):
-                print(file_name + " is a dta file")
-                print("Keeping " + file_name)
-
-            elif file_name.endswith(".log"):
-                print(file_name + " is a log file")
-                print("Keeping " + file_name)
-
-            elif file_name.endswith(".zip"):
-                print(file_name + " is a zip file")
-                if os.path.dirname(directory_path) == "codebooks":
+    if directory_name == "codebooks":
+        if re.fullmatch("[0-9]{4}", directory_name):
+            for file_name in file_names_list:
+                file_path = os.path.join(directory_path, file_name)
+                print("Checking if " + file_name + " is a zip file")
+                if file_name.endswith(".zip"):
+                    print(file_name + " is a zip file")
                     print("Deleting " + file_name)
-                else:
-                    print("Keeping " + file_name)
+                    os.remove(file_path)
 
-            elif file_name.endswith(".do"):
-                print(file_name + " is a do file")
-                print("Deleting " + file_name)
-                os.remove(file_path)
+    if directory_name == "do_files":
+        if re.fullmatch("[0-9]{4}", directory_name):
+            for file_name in file_names_list:
+                file_path = os.path.join(directory_path, file_name)
+                print("Checking if " + file_name + " is a do file")
+                if file_name.endswith(".do"):
+                    print(file_name + " is a do file")
+                    print("Deleting " + file_name)
+                    os.remove(file_path)
 
-            elif file_name.endswith(".csv"):
-                print(file_name + " is a csv file")
-                print("Deleting " + file_name)
-                os.remove(file_path)
-
-            # elif file_name.endswith(".html"):
-            #     print(file_name + " is an html file")
-            #     print("Deleting " + file_name)
-            #     os.remove(file_path)
+    if directory_name == "raw_data":
+        if re.fullmatch("[0-9]{4}", directory_name):
+            for file_name in file_names_list:
+                file_path = os.path.join(directory_path, file_name)
+                print("Checking if " + file_name + " is a csv file")
+                if file_name.endswith(".csv"):
+                    print(file_name + " is a csv file")
+                    print("Deleting " + file_name)
+                    os.remove(file_path)
